# specs/001-hugo-matter-cms/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Hugo Matters API
  version: 0.1.0
  description: |
    Local ApiService contract for Hugo Matters CMS v1.
    Web (Blazor Server) is the primary client via Aspire service discovery.
servers:
  - url: https://apiservice
    description: Aspire service discovery name (scheme may resolve to http in local dev)
...
paths:
...
  /api/session:
    get:
      tags: [Session]
      summary: Get active editing session for the connected site
      responses:
        "200":
          description: Active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditingSession"
        "404":
          description: No active session
    post:
      tags: [Session]
      summary: Start editing session (branch + PR). Fails if one already active.
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditingSession"
        "409":
          description: Active session already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditingSession"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/session/save:
    post:
      tags: [Session]
      summary: Persist buffered content changes as commit(s) on the session branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRequest"
      responses:
        "200":
          description: Save succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/session/publish:
    post:
      tags: [Session]
      summary: Merge PR into default branch, delete session branch, end session
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
        "400":
          description: Blocked (no changes, unsaved edits) or failed merge with explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
        "401":
          $ref: "#/components/responses/Error"
  /api/session/discard:
    post:
      tags: [Session]
      summary: Close PR without merge, delete branch, end session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscardRequest"
      responses:
        "200":
          description: Discarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscardResult"
        "400":
          description: Confirmation required for unsaved edits, or failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscardResult"
...
components:
...
  responses:
    Error:
      description: Safe failure payload (no secrets)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    EditingSession:
      type: object
      required: [id, siteId, branchName, pullRequestNumber, baseBranch, state, hasUnsavedLocalEdits]
      properties:
        id:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        branchName:
          type: string
        pullRequestNumber:
          type: integer
        pullRequestUrl:
          type: string
        baseBranch:
          type: string
        state:
          type: string
          enum: [Active, Publishing, Discarding, Ended]
        hasUnsavedLocalEdits:
          type: boolean
    SaveRequest:
      type: object
      properties:
        commitMessage:
          type: string
          description: Optional; server may supply a default
    SaveResult:
      type: object
      required: [commitSha, hasUnsavedLocalEdits]
      properties:
        commitSha:
          type: string
        hasUnsavedLocalEdits:
          type: boolean
    PublishResult:
      type: object
      required: [outcome, message]
      properties:
        outcome:
          type: string
          enum:
            - Succeeded
            - BlockedNoChanges
            - BlockedUnsavedEdits
            - FailedMerge
            - FailedAuth
        message:
          type: string
        mergeCommitSha:
          type: string
          nullable: true
    DiscardRequest:
      type: object
      required: [confirmDiscardUnsaved]
      properties:
        confirmDiscardUnsaved:
          type: boolean
          description: Must be true when hasUnsavedLocalEdits
    DiscardResult:
      type: object
      required: [outcome, message]
      properties:
        outcome:
          type: string
          enum: [Succeeded, ConfirmationRequired, Failed]
        message:
          type: string
...