# src/HugoMatters.ApiService/Endpoints/SessionEndpoints.cs
using HugoMatters.ApiService.Http;
using HugoMatters.Core.Api;
using HugoMatters.Core.Sessions;

namespace HugoMatters.ApiService.Endpoints;

/// <summary>
/// Editing session lifecycle endpoints.
/// </summary>
public static class SessionEndpoints
{
    /// <summary>
    /// Maps session endpoints.
    /// </summary>
    public static IEndpointRouteBuilder MapSessionEndpoints(this IEndpointRouteBuilder app)
    {
        var group = app.MapGroup("/api/session").WithTags("Session");

        group.MapGet("/", GetSessionAsync)
            .WithName("GetSession");

        group.MapGet("/resumable", ListResumableSessionsAsync)
            .WithName("ListResumableSessions");

        group.MapPost("/", StartSessionAsync)
            .WithName("StartSession");

        group.MapPost("/resume", ResumeSessionAsync)
            .WithName("ResumeSession");

        group.MapPost("/save", SaveAsync)
            .WithName("SaveSession");

        group.MapPost("/publish", PublishAsync)
            .WithName("PublishSession");

        group.MapPost("/discard", DiscardAsync)
            .WithName("DiscardSession");

        group.MapPost("/leave", LeaveAsync)
            .WithName("LeaveSession");

        return app;
    }

    private static async Task<IResult> GetSessionAsync(
        SessionService sessionService,
        CancellationToken cancellationToken)
    {
        var session = await sessionService.GetActiveSessionAsync(cancellationToken);
        return session is null ? Results.NotFound() : Results.Ok(session);
    }

    private static async Task<IResult> ListResumableSessionsAsync(
        SessionService sessionService,
        CancellationToken cancellationToken)
    {
        try
        {
            var items = await sessionService.ListResumableSessionsAsync(cancellationToken);
            return Results.Ok(items);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status400BadRequest);
        }
        catch (UnauthorizedAccessException ex)
        {
            return ApiResults.Error("unauthorized", ex.Message, StatusCodes.Status401Unauthorized);
        }
    }

    private static async Task<IResult> StartSessionAsync(
        SessionService sessionService,
        CancellationToken cancellationToken)
    {
        try
        {
            var session = await sessionService.StartSessionAsync(cancellationToken);
            return Results.Created("/api/session", session);
        }
        catch (InvalidOperationException ex) when (ex.Message.Contains("already exists", StringComparison.OrdinalIgnoreCase))
        {
            var existing = await sessionService.GetActiveSessionAsync(cancellationToken);
            return Results.Conflict(existing);
        }
        catch (InvalidOperationException ex) when (ex.Message.Contains("Pull requests are disabled", StringComparison.OrdinalIgnoreCase)
            || ex.Message.Contains("Pull requests are unavailable", StringComparison.OrdinalIgnoreCase))
        {
            return ApiResults.Error("pull_requests_disabled", ex.Message, StatusCodes.Status400BadRequest);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status400BadRequest);
        }
        catch (UnauthorizedAccessException ex)
        {
            return ApiResults.Error("unauthorized", ex.Message, StatusCodes.Status401Unauthorized);
        }
        catch (Octokit.NotFoundException ex)
        {
            return ApiResults.Error(
                "github_not_found",
                string.IsNullOrWhiteSpace(ex.Message) ? "GitHub resource was not found." : ex.Message,
                StatusCodes.Status404NotFound);
        }
        catch (Octokit.ApiException ex) when ((int)ex.StatusCode is >= 400 and < 500)
        {
            return ApiResults.Error(
                "github_request_failed",
                string.IsNullOrWhiteSpace(ex.Message) ? "GitHub rejected the session request." : ex.Message,
                (int)ex.StatusCode);
        }
    }

    private static async Task<IResult> ResumeSessionAsync(
        ResumeSessionRequest request,
        SessionService sessionService,
        CancellationToken cancellationToken)
    {
        try
        {
            var session = await sessionService.ResumeSessionAsync(request.PullRequestNumber, cancellationToken);
            return Results.Ok(session);
        }
        catch (InvalidOperationException ex) when (ex.Message.Contains("already exists", StringComparison.OrdinalIgnoreCase))
        {
            var existing = await sessionService.GetActiveSessionAsync(cancellationToken);
            return Results.Conflict(existing);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status400BadRequest);
        }
        catch (UnauthorizedAccessException ex)
        {
            return ApiResults.Error("unauthorized", ex.Message, StatusCodes.Status401Unauthorized);
        }
        catch (Octokit.NotFoundException)
        {
            return ApiResults.Error(
                "session_branch_missing",
                "The session branch for that pull request no longer exists on GitHub.",
                StatusCodes.Status409Conflict);
        }
    }

    private static async Task<IResult> SaveAsync(
        SaveRequest? request,
        SaveService saveService,
        CancellationToken cancellationToken)
    {
        try
        {
            var result = await saveService.SaveAsync(request, cancellationToken);
            return Results.Ok(result);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status409Conflict);
        }
        catch (UnauthorizedAccessException ex)
        {
            return ApiResults.Error("unauthorized", ex.Message, StatusCodes.Status401Unauthorized);
        }
    }

    private static async Task<IResult> PublishAsync(
        PublishService publishService,
        CancellationToken cancellationToken)
    {
        try
        {
            var result = await publishService.PublishAsync(cancellationToken);
            if (result.Outcome == HugoMatters.Core.Models.PublishOutcome.Succeeded)
            {
                return Results.Ok(result);
            }

            return Results.Json(result, statusCode: StatusCodes.Status400BadRequest);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status400BadRequest);
        }
        catch (UnauthorizedAccessException ex)
        {
            return ApiResults.Error("unauthorized", ex.Message, StatusCodes.Status401Unauthorized);
        }
    }

    private static async Task<IResult> DiscardAsync(
        DiscardRequest request,
        DiscardService discardService,
        CancellationToken cancellationToken)
    {
        try
        {
            var result = await discardService.DiscardAsync(request, cancellationToken);
            if (result.Outcome == HugoMatters.Core.Models.DiscardOutcome.Succeeded)
            {
                return Results.Ok(result);
            }

            return Results.Json(result, statusCode: StatusCodes.Status400BadRequest);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status400BadRequest);
        }
    }

    private static async Task<IResult> LeaveAsync(
        LeaveSessionRequest request,
        LeaveSessionService leaveSessionService,
        CancellationToken cancellationToken)
    {
        try
        {
            var result = await leaveSessionService.LeaveAsync(request, cancellationToken);
            if (result.Outcome == HugoMatters.Core.Models.LeaveSessionOutcome.Succeeded)
            {
                return Results.Ok(result);
            }

            return Results.Json(result, statusCode: StatusCodes.Status400BadRequest);
        }
        catch (InvalidOperationException ex)
        {
            return ApiResults.Error("invalid_state", ex.Message, StatusCodes.Status400BadRequest);
        }
    }
}